third_party/WebKit/LayoutTests/http/tests/serviceworker/update-no-controllee.https-expected.txt

This is a testharness.js-based test.
FAIL Verify multiple updates from service workers without controllees do not resolve immediately assert_equals: update should not have succeeded expected "failure" but got "success"
Harness: the test ran to completion.
